requirements.txt
Flask==3.1.0
flask-cors==5.0.0
requests==2.32.4
google-api-python-client==2.176.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-auth==2.40.3
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.1.1

